requirements.txt
discord.py
python-dotenv
PyNaCl
yt-dlp
pytube
discord-py-interactions
asyncio